editor: add 'revert and close editor' command
The pull-request moves the `revert and close editor` command out of the
plugin-system so it can be used by end-users as well as for vscode
plugins.
