+++ packages/monaco/src/browser/monaco-command.ts
@@ -17,8 +17,8 @@
 import { injectable, inject, optional } from '@theia/core/shared/inversify';
 import { Position, Location } from '@theia/core/shared/vscode-languageserver-types';
 import { CommandContribution, CommandRegistry, CommandHandler } from '@theia/core/lib/common/command';
-import { CommonCommands, QuickInputService } from '@theia/core/lib/browser';
-import { EditorCommands } from '@theia/editor/lib/browser';
+import { CommonCommands, QuickInputService, ApplicationShell } from '@theia/core/lib/browser';
+import { EditorCommands, EditorManager } from '@theia/editor/lib/browser';
 import { MonacoEditor } from './monaco-editor';
 import { MonacoCommandRegistry, MonacoEditorCommandHandler } from './monaco-command-registry';
 import { MonacoEditorService } from './monaco-editor-service';
@@ -68,6 +68,12 @@ export class MonacoEditorCommandHandlers implements CommandContribution {
 
     @inject(monaco.contextKeyService.ContextKeyService)
     protected readonly contextKeyService: monaco.contextKeyService.ContextKeyService;
+
+    @inject(ApplicationShell)
+    protected readonly shell: ApplicationShell;
+
+    @inject(EditorManager)
+    protected editorManager: EditorManager;
 
     registerCommands(): void {
         this.registerMonacoCommands();
@@ -185,6 +191,7 @@ export class MonacoEditorCommandHandlers implements CommandContribution {
         this.monacoCommandRegistry.registerHandler(EditorCommands.CONFIG_EOL.id, this.newConfigEolHandler());
         this.monacoCommandRegistry.registerHandler(EditorCommands.INDENT_USING_SPACES.id, this.newConfigTabSizeHandler(true));
         this.monacoCommandRegistry.registerHandler(EditorCommands.INDENT_USING_TABS.id, this.newConfigTabSizeHandler(false));
+        this.monacoCommandRegistry.registerHandler(EditorCommands.REVERT_AND_CLOSE.id, this.newRevertAndCloseActiveEditorHandler());
     }
 
     protected newShowReferenceHandler(): MonacoEditorCommandHandler {
@@ -222,10 +229,11 @@ export class MonacoEditorCommandHandlers implements CommandContribution {
 
     protected configureEol(editor: MonacoEditor): void {
         const items = ['LF', 'CRLF'].map(lineEnding =>
-        ({
-            label: lineEnding,
-            execute: () => this.setEol(editor, lineEnding)
-        }));
+            ({
+                label: lineEnding,
+                execute: () => this.setEol(editor, lineEnding)
+            })
+        );
         this.quickInputService?.showQuickPick(items, { placeholder: 'Select End of Line Sequence' });
     }
 
@@ -251,15 +259,37 @@ export class MonacoEditorCommandHandlers implements CommandContribution {
             const { tabSize } = model.getOptions();
             const sizes = Array.from(Array(8), (_, x) => x + 1);
             const tabSizeOptions = sizes.map(size =>
-            ({
-                label: size === tabSize ? `${size}   Configured Tab Size` : size.toString(),
-                // eslint-disable-next-line @typescript-eslint/no-explicit-any
-                execute: (quickPick: any, lookFor: string) => model.updateOptions({
-                    tabSize: size || tabSize,
-                    insertSpaces: useSpaces
+                ({
+                    label: size === tabSize ? `${size}   Configured Tab Size` : size.toString(),
+                    // eslint-disable-next-line @typescript-eslint/no-explicit-any
+                    execute: (quickPick: any, lookFor: string) => model.updateOptions({
+                        tabSize: size || tabSize,
+                        insertSpaces: useSpaces
+                    })
                 })
-            }));
+            );
             this.quickInputService?.showQuickPick(tabSizeOptions, { placeholder: 'Select Tab Size for Current File' });
+        }
+    }
+
+    protected newRevertAndCloseActiveEditorHandler(): MonacoEditorCommandHandler {
+        return {
+            execute: async () => this.revertAndCloseActiveEditor()
+        };
+    }
+
+    protected async revertAndCloseActiveEditor(): Promise<void> {
+        const editor = this.editorManager.currentEditor;
+        if (editor) {
+            const monacoEditor = MonacoEditor.getCurrent(this.editorManager);
+            if (monacoEditor) {
+                try {
+                    await monacoEditor.document.revert();
+                    editor.close();
+                } catch (error) {
+                    await this.shell.closeWidget(editor.id, { save: false });
+                }
+            }
         }
     }
 }
